app/requirements.txt
django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytest-django==4.8.0
pytest==8.0.0
celery[redis]==5.3.6
django-redis==5.4.0
django-cachalot==2.6.2
requests==2.31.0
pytest-cov==4.1.0
flake8==7.0.0
black==24.2.0
isort==5.13.2
gunicorn==21.2.0
django-cors-headers==4.3.1
python-telegram-bot==13.15
stripe>=8.0.0
